gcp/requirements.txt
tensorflow==2.18.0
google-cloud-storage==1.16.1
pillow==11.1.0